+++ RELEASE_NOTES.md
@@ -9,6 +9,9 @@ If you are enjoying FreeAIr to the enough level to donate, there are many [small
 ## 4.0.0
 
 - Added [in situ chat](https://raw.githubusercontent.com/lsoft/FreeAIr/main/in_situ_chat.gif) mode.
+- BREAKING CHANGE: changed MCP servers json format in FreeAIr options json file.
+- Added easy way to enable Microsoft MDSN MCP server.
+- Added asking permission from the user to call any MCP tools LLM requested.
 - Fixed various bugs.
 
 ## 3.7.1