+++ RELEASE_NOTES.md
@@ -6,7 +6,7 @@ Please report any bugs to the [github repo](https://github.com/lsoft/FreeAIr).
 
 If you are enjoying FreeAIr to the enough level to donate, there are many [small cancer patients](https://advitausa.org/au/index.php/donate/) that need your help. Please provide your help them!
 
-## 4.0.0
+## 4.0.2
 
 - Added [in situ chat](https://raw.githubusercontent.com/lsoft/FreeAIr/main/in_situ_chat.gif) mode.
 - BREAKING CHANGE: changed MCP servers json format in FreeAIr options json file.